requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==4.2.6
django-filter==23.3
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Markdown==3.5
PyJWT==2.8.0
pytz==2023.3.post1
redis==5.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
